// src/models/NameModels.ts
export enum NameModels {
  antecedent = 'antecedent',
  compteBloque = 'compteBloque',
  consultation = 'consultation',
  dossierPatient = 'dossierPatient',
  employer = 'employer',
  historiqueMessage = 'historiqueMessage',
   infoClinique = 'infoClinique',
  medecin = 'medecin',
  message = 'message',
  ordonnance = 'ordonnance',
  patient = 'patient',
  payement = 'payement',
  personnel = 'personnel',
  poste = 'poste',
  rendezVous = 'rendezVous',
  salle = 'salle',
  tache = 'tache',
  typeDeSalle = 'typeDeSalle',
  typeRendezVous = 'typeRendezVous',
}